wip and roadtrip method to service and add route for road trip, edit forecast service and refactor naming convention for dev empathy
