+++ config/routes.rb
@@ -6,6 +6,7 @@ Rails.application.routes.draw do
       get '/backgrounds', to: 'backgrounds#show'
       post '/users', to: 'users#registration'
       post '/sessions', to: 'sessions#login'
+      post '/road_trip', to: 'road_trips#create'
     end
   end
 end
